requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
Flask-Login>=0.6
Flask-Migrate>=4.0
